add a utf8 decoder helper
xslate doesn't provide any way to mark code inside template files as
utf8, so strings in that code end up as undecoded strings, even if the
rest of the template file itself is decoded. this provides a helper
function you can use in your templates to explicitly decode the strings.
if this ever gets fixed in xslate, i'll be able to just conditionally
turn this helper function into a noop, so it shouldn't really affect
forwards compatibility.